Move AsyncSpec into test directory.
